+++ packages/nimble-components/src/table/index.ts
@@ -1,17 +1,17 @@
 import { observable } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import {
-    ColumnDef,
-    TableState,
-    Updater,
-    Table as TanstackTable,
-    createTable,
-    getCoreRowModel,
-    TableOptionsResolved
+    ColumnDef as TanStackColumnDef,
+    TableState as TanStackTableState,
+    Updater as TanStackUpdater,
+    Table as TanStackTable,
+    createTable as tanStackCreateTable,
+    getCoreRowModel as tanStackGetCoreRowModel,
+    TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
 import { styles } from './styles';
 import { template } from './template';
-import type { TableData } from './types';
+import type { TableRecord, TableFieldValue } from './types';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -23,39 +23,35 @@ declare global {
  * A nimble-styled table.
  */
 export class Table<
-    TData extends TableData = {
-        [key: string]: string | number | boolean | Date | null | undefined
-    }
+    TData extends TableRecord = TableRecord
 > extends FoundationElement {
     @observable
     public data: TData[] = [];
 
-    // TODO: Temporarily assume that the data in the table can be represented as strings.
+    // TODO: Temporarily expose the table data as arrays of strings.
     @observable
     public tableData: string[][] = [];
 
+    // TODO: Temporarily expose the column headers as a string array.
     @observable
     public columnHeaders: string[] = [];
 
-    private readonly table: TanstackTable<TData>;
-    private options: TableOptionsResolved<TData>;
+    private readonly table: TanStackTable<TData>;
+    private options: TanStackTableOptionsResolved<TData>;
     private readonly tableInitialized: boolean = false;
 
     public constructor() {
         super();
-        const initialData = this.data;
         this.options = {
-            get data(): TData[] {
-                return initialData;
-            },
-            onStateChange: (_: Updater<TableState>) => {},
-            getCoreRowModel: getCoreRowModel(),
+            data: [],
+            onStateChange: (_: TanStackUpdater<TanStackTableState>) => {},
+            getCoreRowModel: tanStackGetCoreRowModel(),
             columns: [],
             state: {},
             renderFallbackValue: null,
             autoResetAll: false
         };
-        this.table = createTable(this.options);
+        this.table = tanStackCreateTable(this.options);
         this.tableInitialized = true;
     }
 
@@ -82,7 +78,9 @@ export class Table<
         for (const row of rows) {
             const rowArray: string[] = [];
             for (const cell of row.getVisibleCells()) {
-                rowArray.push(cell.getValue() as string);
+                const cellValue = cell.getValue() as TableFieldValue;
+                const stringValue = cellValue == null ? '' : cellValue.toString();
+                rowArray.push(stringValue);
             }
             tableData.push(rowArray);
         }
@@ -90,21 +88,21 @@ export class Table<
     }
 
     private updateTableOptions(
-        updatedOptions: Partial<TableOptionsResolved<TData>>
+        updatedOptions: Partial<TanStackTableOptionsResolved<TData>>
     ): void {
         this.options = { ...this.options, ...updatedOptions };
         this.update(this.table.initialState);
     }
 
-    private readonly update = (state: TableState): void => {
+    private readonly update = (state: TanStackTableState): void => {
         this.table.setOptions(prev => ({
             ...prev,
             ...this.options,
             state,
             onStateChange: (updater: unknown) => {
                 const updatedState = typeof updater === 'function'
-                    ? (updater(state) as TableState)
-                    : (updater as TableState);
+                    ? (updater(state) as TanStackTableState)
+                    : (updater as TanStackTableState);
                 this.update(updatedState);
             }
         }));
@@ -120,7 +118,7 @@ export class Table<
         const firstItem = this.data[0]!;
         const keys = Object.keys(firstItem);
         const generatedColumns = keys.map(key => {
-            const columnDef: ColumnDef<TData> = {
+            const columnDef: TanStackColumnDef<TData> = {
                 id: key,
                 accessorKey: key,
                 header: key