+++ packages/nimble-components/src/table/tests/table.stories.ts
@@ -3,13 +3,36 @@ import type { Meta, StoryObj } from '@storybook/html';
 import { withXD } from 'storybook-addon-xd-designs';
 import { createUserSelectedThemeStory } from '../../utilities/tests/storybook';
 import '../../all-components';
+import { ExampleDataType } from './types';
 import { bodyFont } from '../../theme-provider/design-tokens';
 
 interface TableArgs {
-    data: { myStr: string, myNum: number, myBool: boolean }[];
+    data: ExampleDataType;
 }
 
+const simpleData = [
+    {
+        myStr: 'my first row',
+        myNum: 5,
+        myBool: true
+    },
+    {
+        myStr: 'my second row',
+        myNum: 15,
+        myBool: false
+    }
+] as const;
+
+const dataSets = {
+    [ExampleDataType.simpleData]: simpleData
+} as const;
+
 const overviewText = 'The `nimble-table` is a component that offers a way to render tabular data in a variety of ways in each column.';
+
+const dataDescription = `\`data\` is a property that is an array of records. A record provides the data that backs a single row in the table.
+Each record is made up of fields, which are key/value pairs. The key in each pair must be of type \`string\`, which is defined by the type
+\`TableFieldName\`. The value in each pair must be of type \`string\`, \`number\`, \`boolean\`, \`Date\`, \`null\`, or \`undefined\`,
+which is defined by the type \`TableFieldValue\`.`;
 
 const metadata: Meta<TableArgs> = {
     title: 'Table',
@@ -34,7 +57,7 @@ const metadata: Meta<TableArgs> = {
             WARNING - The table is still in development and considered
             experimental. It is not recommended for application use.
         </div>
-        <nimble-table :data=${x => x.data}>
+        <nimble-table :data=${x => dataSets[x.data]}>
         </nimble-table>
         <style class="code-hide">
             #usage-warning {
@@ -43,19 +66,23 @@ const metadata: Meta<TableArgs> = {
             }
         </style>
     `),
-    args: {
-        data: [
-            {
-                myStr: 'my first row',
-                myNum: 5,
-                myBool: true
+    argTypes: {
+        data: {
+            description: dataDescription,
+            table: {
+                defaultValue: { summary: '[]' }
             },
-            {
-                myStr: 'my second row',
-                myNum: 15,
-                myBool: false
+            options: [ExampleDataType.simpleData],
+            control: {
+                type: 'radio',
+                labels: {
+                    [ExampleDataType.simpleData]: 'Simple data'
+                }
             }
-        ]
+        }
+    },
+    args: {
+        data: ExampleDataType.simpleData
     }
 };