+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Mon, 05 Dec 2022 19:44:44 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Dec 2022 18:13:52 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 14.0.12
+
+Mon, 12 Dec 2022 18:13:52 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.6
+
+## 14.0.11
+
+Fri, 09 Dec 2022 23:00:24 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.5
+
+## 14.0.10
+
+Fri, 09 Dec 2022 11:25:00 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.4
+
+## 14.0.9
+
+Wed, 07 Dec 2022 13:10:43 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v15.5.3
 
 ## 14.0.8