+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Mon, 05 Dec 2022 19:44:44 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Dec 2022 18:13:52 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 15.5.6
+
+Mon, 12 Dec 2022 18:13:52 GMT
+
+### Patches
+
+- Created render queue for wafer map component input changes ([ni/nimble@e70046b](https://github.com/ni/nimble/commit/e70046b7a09cd85253303ba5ec4bf2ab314c9c6d))
+
+## 15.5.5
+
+Fri, 09 Dec 2022 23:00:24 GMT
+
+### Patches
+
+- Basic TanStack integration with nimble-table ([ni/nimble@a4ce35d](https://github.com/ni/nimble/commit/a4ce35df8efd5b806cc0335b30a2c4764d470a8c))
+
+## 15.5.4
+
+Fri, 09 Dec 2022 11:25:00 GMT
+
+### Patches
+
+- Finished first iteration of the waferMap component and its storybook ([ni/nimble@001156d](https://github.com/ni/nimble/commit/001156d6c85f250a934ed8819b6a0650fe26eb84))
+
+## 15.5.3
+
+Wed, 07 Dec 2022 13:10:43 GMT
+
+### Patches
+
+- Created data manager module, prerendering functionality and modified some interface inputs ([ni/nimble@b7e946b](https://github.com/ni/nimble/commit/b7e946be21b68e6b871096ec587defd543d6ce32))
 
 ## 15.5.2